third_party/blink/web_tests/platform/generic/fast/forms/text/text-update-datalist-while-focused-expected.txt

We should restore selection after datalist availability change.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input1.value is "abc"
PASS successfullyParsed is true

TEST COMPLETE
